add component documentation
