+++ src/components/BlockButton/__tests__/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { shallow } from "enzyme";
+import BlockButton from "../index";
+
+describe("BlockButton", () => {
+  it("should render successfully", () => {
+    //   const component = shallow(<BlockButton />);
+    //   expect(component).toMatchSnapshot();
+  });
+});
