+++ src/components/MarkButton/__tests__/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { shallow } from "enzyme";
+import MarkButton from "../index";
+
+describe("MarkButton", () => {
+  it("should render successfully", () => {
+  //   const component = shallow(<MarkButton />);
+  //   expect(component).toMatchSnapshot();
+  });
+});
